refactor: #151 기존 GPSLog 수집에 대한 인터페이스화 Controller 에서 Service 로 변경
