+++ src/main/java/org/thisway/log/consumer/GpsLogConsumer.java
@@ -6,18 +6,18 @@ import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Component;
 import org.thisway.config.ampq.RabbitMQConfig;
 import org.thisway.log.dto.request.gpsLog.GpsLogRequest;
-import org.thisway.log.service.LogService;
+import org.thisway.log.service.GpsLogSaveService;
 
 @Component
 @RequiredArgsConstructor
 @Slf4j
 public class GpsLogConsumer {
 
-    private final LogService logService;
+    private final GpsLogSaveService gpsLogSaveService;
 
     @RabbitListener(queues = RabbitMQConfig.GPS_LOG_QUEUE)
     public void receiveGpsLog(GpsLogRequest request) {
         log.debug("Received GPS log: {}", request);
-        logService.saveGpsLog(request);
+        gpsLogSaveService.saveGpsLog(request);
     }
 }
