+++ src/main/java/org/thisway/log/service/LogService.java
@@ -1,8 +1,5 @@
 package org.thisway.log.service;
 
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -13,11 +10,8 @@ import org.thisway.emulator.entity.Emulator;
 import org.thisway.emulator.repository.EmulatorRepository;
 import org.thisway.log.converter.LogDataConverter;
 import org.thisway.log.domain.GeofenceLogData;
-import org.thisway.log.domain.GpsLogData;
 import org.thisway.log.domain.PowerLogData;
 import org.thisway.log.dto.request.geofenceLog.GeofenceLogRequest;
-import org.thisway.log.dto.request.gpsLog.GpsLogEntry;
-import org.thisway.log.dto.request.gpsLog.GpsLogRequest;
 import org.thisway.log.dto.request.powerLog.PowerLogRequest;
 import org.thisway.log.repository.LogRepository;
 import org.thisway.vehicle.repository.VehicleRepository;
@@ -47,7 +41,7 @@ public class LogService {
             logRepository.savePowerLog(powerLogData);
             vehicleRepository.findById(vehicleId).ifPresent(vehicle -> {
                 vehicle.updatePowerOn(true);
-                
+
                 vehicleRepository.save(vehicle);
             });
             log.info("시동 ON 정보 로그 저장: MDN={}, onTime={}", request.mdn(), request.onTime());
@@ -74,38 +68,6 @@ public class LogService {
         }
 
         log.info("시동 정보 로그 저장 완료: MDN={}", request.mdn());
-    }
-
-    public void saveGpsLog(GpsLogRequest request) {
-        log.info("주기 정보 로그 수신: MDN={}, 항목 수={}, 시간={}", request.mdn(), request.cCnt(), request.oTime());
-
-        String mdn = request.mdn();
-        Long vehicleId = getVehicleIdByMdn(mdn);
-
-        List<GpsLogData> gpsLogDataList = new ArrayList<>();
-
-        LocalDateTime occurredTime;
-        try {
-            if (request.oTime().length() == 14) {
-                occurredTime = converter.convertDateTimeWithSec(request.oTime());
-                log.info("초 단위 시간 형식 감지: {}", request.oTime());
-            } else {
-                occurredTime = converter.convertDateTime(request.oTime());
-                log.info("분 단위 시간 형식 감지: {}", request.oTime());
-            }
-        } catch (Exception e) {
-            log.error("시간 형식 변환 오류: {}, 오류 메시지: {}", request.oTime(), e.getMessage());
-            throw new CustomException(ErrorCode.SERVER_ERROR);
-        }
-
-        for (GpsLogEntry entry : request.cList()) {
-            GpsLogData gpsLogData = GpsLogData.from(entry, mdn, vehicleId, occurredTime, converter);
-            gpsLogDataList.add(gpsLogData);
-        }
-
-        logRepository.saveGpsLogs(gpsLogDataList);
-
-        log.info("주기 정보 로그 저장 완료: MDN={}, 항목 수={}", request.mdn(), gpsLogDataList.size());
     }
 
     public void saveGeofenceLog(GeofenceLogRequest request) {
