+++ src/test/resources/application.yml
@@ -41,3 +41,5 @@ security:
   jwt:
     secret-key: szU3FELDRTw9sJzzP1cWuo/xUzIjSvMGGSWKlMCgku/vZwsSmi8BHsH/rnsAyTzTAQ2IGS0mgr/oxi4WsHMhcQ==
     access-token-validity-ms: 360000
+
+gps-log-collect-mode: direct
